Added "The Art of Doing Science and Engineering"
While this book is not about AI specifically, Hamming covers what he thought AI would evolve into (which it kind of did) in chapters 6, 7, and 8. Even with what we are now capable of doing, he doesn’t think we need fewer programmers, just smarter ones. And that we should focus on what problems to solve, and why.