Added: "Deep Learning"
The most complete and in-depth book on Deep Learning. It’s dense but thorough, covering the math, statistics, and core algorithms that form the foundation of modern AI. It helped me build a clearer mental model of the tools used in research and implementation. I posted a video with my raw thoughts on the book: https://youtu.be/DtOQ6zcgqfY?si=PHYL4q5rMaCF-gEW